Update documentation to use API token instead of username and password
Given the recent [blog post](https://blog.pinboard.in/2014/10/holy_war_on_sites_that_demand_pinboard_passwords/)
on avoiding usernames and passwords to authenticate with Pinboard, I was
going to rewrite the `Client` class to accept api tokens.

After digging into the source, I noticed that this functionality was
already there, but the readme did not indicate this was possible.